MergeStack: convert to functional component.
